+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 - `network::client::tcp::TcpStream::connect_timeout` method.
 - Now `test` attribute supports adding a compile time conditional expression for
   the `should_panic` argument, i.e. `#[tarantool::test(should_panic = cfg!(debug_assertions))]`
+- `anyhow` feature that adds `IntoBoxError` impl for `anyhow::Error` so `anyhow::Result` could be
+  returned from the stored proc.
 
 ### Changed